requirements.txt
pandas
langchain
scipy
numpy
pypdf
cryptography
Pillow
streamlit-chat
streamlit
